--- .github/workflows/pgtrio.yml
@@ -1,51 +1,0 @@
-name: Unit Tests
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.8", "3.9", "3.10"]
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Extract branch name
-      shell: bash
-      run: echo "##[set-output name=branch;]$(echo ${GITHUB_REF#refs/heads/})"
-      id: extract_branch
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v3
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install ubuntu packages
-      run: |
-        sudo apt-get update
-        sudo apt-get install -y postgresql-12 jq
-    - name: Install dependencies
-      run: |
-        python -m pip install -r requirements.txt
-        python -m pip install pytest-rerunfailures==10.2
-    - name: Test with pytest
-      run: |
-        coverage run --source=pgtrio -m pytest -v --reruns=5
-    - name: Create coverage badge
-      run: |
-        COVERAGE=$(coverage json -o- | jq '.totals.percent_covered | round')
-        coverage-badge >coverage.svg
-    - name: Upload badge to gist
-      uses: exuanbo/actions-deploy-gist@v1
-      with:
-        token: ${{ secrets.BADGE_GIST_TOKEN }}
-        gist_id: 31aafc23e3119da1d39e1b9aaf5a43fd
-        gist_description: "pgtrio ${{ steps.extract_branch.outputs.branch }}"
-        gist_file_name: pgtrio-${{ steps.extract_branch.outputs.branch }}-coverage.svg
-        file_path: coverage.svg
-
