--- pgtrio/__init__.py
@@ -1,9 +1,0 @@
-from ._connection import connect
-from ._pool import create_pool
-from ._utils import PgProtocolFormat, PgIsolationLevel, PgReadWriteMode
-from ._codecs import Codec
-from ._exceptions import (
-    Warning, Error, InterfaceError, DatabaseError, DataError,
-    OperationalError, IntegrityError, InternalError, ProgrammingError,
-    NotSupportedError,
-)
