+++ tests/test_connect.py
@@ -1,6 +1,9 @@
-import trio
-import pgtrio
-from unittest.mock import patch, AsyncMock
+# pyright: reportAttributeAccessIssue=none
+from unittest.mock import patch
+
+import anyio
+
+import pganyio
 
 
 class MockConnection:
@@ -9,7 +12,7 @@ class MockConnection:
         self.kwargs = kwargs
 
         self._is_ready = True
-        self._is_ready_cv = trio.Condition()
+        self._is_ready_cv = anyio.Condition()
 
     async def _run(self):
         async with self._is_ready_cv:
@@ -20,32 +23,32 @@ class MockConnection:
 
 
 async def test_connect_simple():
-    with patch('pgtrio._connection.Connection', MockConnection):
-        async with pgtrio.connect('postgresql:///foodb') as conn:
-            assert conn.kwargs.get('username') is None
-            assert conn.kwargs.get('password') is None
-            assert conn.kwargs.get('host') is None
-            assert conn.kwargs.get('port') is None
-            assert conn.args == ('foodb',)
+    with patch("pganyio._connection.Connection", MockConnection):
+        async with pganyio.connect("postgresql:///foodb") as conn:
+            assert conn.kwargs.get("username") is None
+            assert conn.kwargs.get("password") is None
+            assert conn.kwargs.get("host") is None
+            assert conn.kwargs.get("port") is None
+            assert conn.args == ("foodb",)
 
 
 async def test_connect_complex():
-    with patch('pgtrio._connection.Connection', MockConnection):
-        url = 'postgresql://user:pwd@localhost:15432/foodb'
-        async with pgtrio.connect(url) as conn:
-            assert conn.kwargs.get('username') == 'user'
-            assert conn.kwargs.get('password') == 'pwd'
-            assert conn.kwargs.get('host') == 'localhost'
-            assert conn.kwargs.get('port') == 15432
-            assert conn.args == ('foodb',)
+    with patch("pganyio._connection.Connection", MockConnection):
+        url = "postgresql://user:pwd@localhost:15432/foodb"
+        async with pganyio.connect(url) as conn:
+            assert conn.kwargs.get("username") == "user"
+            assert conn.kwargs.get("password") == "pwd"
+            assert conn.kwargs.get("host") == "localhost"
+            assert conn.kwargs.get("port") == 15432
+            assert conn.args == ("foodb",)
 
 
 async def test_connect_no_url():
-    with patch('pgtrio._connection.Connection', MockConnection):
-        url = 'postgresql://user:pwd@localhost:15432/foodb'
-        async with pgtrio.connect('foodb') as conn:
-            assert conn.kwargs.get('username') is None
-            assert conn.kwargs.get('password') is None
-            assert conn.kwargs.get('host') is None
-            assert conn.kwargs.get('port') is None
-            assert conn.args == ('foodb',)
+    with patch("pganyio._connection.Connection", MockConnection):
+        _url = "postgresql://user:pwd@localhost:15432/foodb"
+        async with pganyio.connect("foodb") as conn:
+            assert conn.kwargs.get("username") is None
+            assert conn.kwargs.get("password") is None
+            assert conn.kwargs.get("host") is None
+            assert conn.kwargs.get("port") is None
+            assert conn.args == ("foodb",)
