+++ tests/test_cursor.py
@@ -1,21 +1,22 @@
 from pytest import raises
-from pgtrio import ProgrammingError
+
+from pganyio import ProgrammingError
 
 
 async def test_no_transaction(conn):
-    await conn.execute('create table foobar (foo int)')
+    await conn.execute("create table foobar (foo int)")
 
     with raises(ProgrammingError):
-        cur = await conn.cursor('select foo from foobar order by foo')
+        _cur = await conn.cursor("select foo from foobar order by foo")
 
 
 async def test_simple(conn):
-    await conn.execute('create table foobar (foo int)')
+    await conn.execute("create table foobar (foo int)")
     for i in range(5):
-        await conn.execute('insert into foobar (foo) values ($1)', i)
+        await conn.execute("insert into foobar (foo) values ($1)", i)
 
     async with conn.transaction():
-        cur = await conn.cursor('select foo from foobar order by foo')
+        cur = await conn.cursor("select foo from foobar order by foo")
 
         result = await cur.fetch_row()
         assert result == (0,)
@@ -37,12 +38,12 @@ async def test_simple(conn):
 
 
 async def test_iter(conn):
-    await conn.execute('create table foobar (foo int)')
+    await conn.execute("create table foobar (foo int)")
     for i in range(100):
-        await conn.execute('insert into foobar (foo) values ($1)', i * 100)
+        await conn.execute("insert into foobar (foo) values ($1)", i * 100)
 
     async with conn.transaction():
-        cur = await conn.cursor('select * from foobar order by foo')
+        cur = await conn.cursor("select * from foobar order by foo")
         await cur.forward(10)
 
         i = 10
@@ -52,14 +53,14 @@ async def test_iter(conn):
 
 
 async def test_prepared_stmt_no_transaction(conn):
-    await conn.execute('create table foobar (foo int)')
-    stmt = await conn.prepare('select * from foobar')
+    await conn.execute("create table foobar (foo int)")
+    stmt = await conn.prepare("select * from foobar")
     with raises(ProgrammingError):
-        cur = await stmt.cursor()
+        _cur = await stmt.cursor()
 
 
 async def test_prepared_stmt(conn):
-    stmt = await conn.prepare('select generate_series(0, 3)')
+    stmt = await conn.prepare("select generate_series(0, 3)")
     async with conn.transaction():
         cur = await stmt.cursor()
         assert await cur.fetch_row() == (0,)
