+++ tests/test_prepared_stmt.py
@@ -1,41 +1,42 @@
-import trio
 from pytest import raises
-from pgtrio import InterfaceError, ProgrammingError
+
+from pganyio import InterfaceError, ProgrammingError
 
 
 async def test_simple(conn):
-    await conn.execute('create table foobar (foo int)')
+    await conn.execute("create table foobar (foo int)")
 
-    stmt = await conn.prepare('insert into foobar (foo) values (10)')
+    stmt = await conn.prepare("insert into foobar (foo) values (10)")
     await stmt.execute()
 
-    results = await conn.execute('select * from foobar')
+    results = await conn.execute("select * from foobar")
     assert results == [(10,)]
 
 
 async def test_params(conn):
-    await conn.execute('create table foobar (foo int)')
+    await conn.execute("create table foobar (foo int)")
 
-    stmt = await conn.prepare('insert into foobar (foo) values ($1)')
+    stmt = await conn.prepare("insert into foobar (foo) values ($1)")
     await stmt.execute(10)
 
-    results = await conn.execute('select * from foobar')
+    results = await conn.execute("select * from foobar")
     assert results == [(10,)]
 
 
 async def test_param_descs(conn):
-    stmt = await conn.prepare('select generate_series(0, $1)')
+    stmt = await conn.prepare("select generate_series(0, $1)")
     param_descs = stmt.parameters
     assert len(param_descs) == 1
-    assert param_descs[0].type == 'int4'
+    assert param_descs[0].type == "int4"
+
 
 async def test_multipart(conn):
-    await conn.execute('create table foobar (foo int)')
-    await conn.execute('insert into foobar (foo) values (10)')
-    await conn.execute('insert into foobar (foo) values (20)')
-    await conn.execute('insert into foobar (foo) values (30)')
+    await conn.execute("create table foobar (foo int)")
+    await conn.execute("insert into foobar (foo) values (10)")
+    await conn.execute("insert into foobar (foo) values (20)")
+    await conn.execute("insert into foobar (foo) values (30)")
 
-    stmt = await conn.prepare('select * from foobar')
+    stmt = await conn.prepare("select * from foobar")
 
     async with conn.transaction():
         results = await stmt.execute(limit=1)
@@ -46,12 +47,12 @@ async def test_multipart(conn):
 
 
 async def test_forward(conn):
-    await conn.execute('create table foobar (foo int)')
-    await conn.execute('insert into foobar (foo) values (10)')
-    await conn.execute('insert into foobar (foo) values (20)')
-    await conn.execute('insert into foobar (foo) values (30)')
+    await conn.execute("create table foobar (foo int)")
+    await conn.execute("insert into foobar (foo) values (10)")
+    await conn.execute("insert into foobar (foo) values (20)")
+    await conn.execute("insert into foobar (foo) values (30)")
 
-    stmt = await conn.prepare('select * from foobar')
+    stmt = await conn.prepare("select * from foobar")
 
     async with conn.transaction():
         results = await stmt.execute(limit=1)
@@ -64,12 +65,12 @@ async def test_forward(conn):
 
 
 async def test_non_matching_transactions(conn):
-    await conn.execute('create table foobar (foo int)')
-    await conn.execute('insert into foobar (foo) values (10)')
-    await conn.execute('insert into foobar (foo) values (20)')
-    await conn.execute('insert into foobar (foo) values (30)')
+    await conn.execute("create table foobar (foo int)")
+    await conn.execute("insert into foobar (foo) values (10)")
+    await conn.execute("insert into foobar (foo) values (20)")
+    await conn.execute("insert into foobar (foo) values (30)")
 
-    stmt = await conn.prepare('select * from foobar')
+    stmt = await conn.prepare("select * from foobar")
 
     async with conn.transaction():
         await stmt.execute(limit=1)
@@ -80,12 +81,12 @@ async def test_non_matching_transactions(conn):
 
 
 async def test_finished(conn):
-    await conn.execute('create table foobar (foo int)')
-    await conn.execute('insert into foobar (foo) values (10)')
-    await conn.execute('insert into foobar (foo) values (20)')
-    await conn.execute('insert into foobar (foo) values (30)')
+    await conn.execute("create table foobar (foo int)")
+    await conn.execute("insert into foobar (foo) values (10)")
+    await conn.execute("insert into foobar (foo) values (20)")
+    await conn.execute("insert into foobar (foo) values (30)")
 
-    stmt = await conn.prepare('select * from foobar')
+    stmt = await conn.prepare("select * from foobar")
     assert not stmt.finished
 
     async with conn.transaction():
@@ -103,12 +104,12 @@ async def test_finished(conn):
 
 
 async def test_multipart_no_transaction1(conn):
-    await conn.execute('create table foobar (foo int)')
-    await conn.execute('insert into foobar (foo) values (10)')
-    await conn.execute('insert into foobar (foo) values (20)')
-    await conn.execute('insert into foobar (foo) values (30)')
+    await conn.execute("create table foobar (foo int)")
+    await conn.execute("insert into foobar (foo) values (10)")
+    await conn.execute("insert into foobar (foo) values (20)")
+    await conn.execute("insert into foobar (foo) values (30)")
 
-    stmt = await conn.prepare('select * from foobar')
+    stmt = await conn.prepare("select * from foobar")
 
     # using limit outside a transaction block is not allowed; because
     # when you want to do exec_continue, postgres will complain that
@@ -118,12 +119,12 @@ async def test_multipart_no_transaction1(conn):
 
 
 async def test_multipart_no_transaction2(conn):
-    await conn.execute('create table foobar (foo int)')
-    await conn.execute('insert into foobar (foo) values (10)')
-    await conn.execute('insert into foobar (foo) values (20)')
-    await conn.execute('insert into foobar (foo) values (30)')
+    await conn.execute("create table foobar (foo int)")
+    await conn.execute("insert into foobar (foo) values (10)")
+    await conn.execute("insert into foobar (foo) values (20)")
+    await conn.execute("insert into foobar (foo) values (30)")
 
-    stmt = await conn.prepare('select * from foobar')
+    stmt = await conn.prepare("select * from foobar")
 
     async with conn.transaction():
         await stmt.execute(limit=1)
